refactor: Layout 폴더 구조 ㅅ정
