+++ src/components/layout/general-layout/GeneralHero.tsx
