+++ src/components/layout/general-layout/index.ts
@@ -1,0 +1,2 @@
+export { default as GeneralLayout } from "./GeneralLayout"
+export * from "./GeneralLayout"
